add Hevy tool
