+++ .gitignore
@@ -1,4 +1,5 @@
 .venv/
+venv/
 secrets/
 __pycache__
 .env
