+++ Dockerfile
@@ -8,20 +8,41 @@ COPY requirements.txt /app
 
 RUN pip install -r requirements.txt
 
-# Install Chromium and dependencies
 RUN apt-get update && apt-get install -y \
     wget \
-    chromium \
-    chromium-driver \
+    gnupg \
     fonts-liberation \
     libasound2 \
+    libatk-bridge2.0-0 \
+    libatk1.0-0 \
+    libcups2 \
+    libdbus-1-3 \
+    libdrm2 \
     libgbm1 \
-    libu2f-udev \
-    libvulkan1 \
+    libglib2.0-0 \
+    libgtk-3-0 \
+    libnspr4 \
+    libnss3 \
+    libpango-1.0-0 \
+    libx11-6 \
+    libxcb1 \
+    libxcomposite1 \
+    libxdamage1 \
+    libxext6 \
+    libxfixes3 \
+    libxkbcommon0 \
+    libxrandr2 \
+    xdg-utils \
+    --no-install-recommends \
     && rm -rf /var/lib/apt/lists/*
 
-# Set Chromium as default browser
-ENV BROWSER="/usr/bin/chromium --no-sandbox"
+RUN wget -q -O - https://dl.google.com/linux/linux_signing_key.pub | gpg --dearmor -o /usr/share/keyrings/google-chrome.gpg \
+    && echo "deb [arch=amd64 signed-by=/usr/share/keyrings/google-chrome.gpg] http://dl.google.com/linux/chrome/deb/ stable main" > /etc/apt/sources.list.d/google-chrome.list \
+    && apt-get update \
+    && apt-get install -y google-chrome-stable \
+    && rm -rf /var/lib/apt/lists/*
+
+RUN google-chrome --version
 
 COPY . /app
