+++ config/.env.exemple
@@ -1,0 +1,2 @@
+HEAVY_USER_NAME=name
+HEAVY_PASSWORD=password
