+++ src/app.py
@@ -1,19 +1,24 @@
 import streamlit as st
-from agents import google_calendar
+from agents import hevy
+
+agent_constructors = {
+    "Hevy": hevy.get_agent,
+}
+
+@st.cache_resource
+def get_agent(agent_name):
+    return agent_constructors[agent_name]()
+    
 
 st.title("TITAN")
 
-# Sidebar to select an agent
-agent_list = {
-    "Google Calendar": google_calendar.get_agent,
-}
+selected_agent_name = st.sidebar.selectbox("Select an agent:", list(agent_constructors.keys()))
 
-selected_agent_name = st.sidebar.selectbox("Select an agent:", list(agent_list.keys()))
+# This will only initialize once per agent selection
+agent = get_agent(selected_agent_name)
 
 user_input = st.text_input("Enter a request:")
 
 if user_input:
-    agent = agent_list[selected_agent_name]()
     response = agent.run(user_input)
-    
     st.write(response.content)
